feat: overhaul styling of register page + components
