+++ Client/src/Components/LoginRegister/ValidationChecklist.tsx
@@ -1,4 +1,4 @@
-import React from 'react';
+import React from "react";
 import { CheckCircleIcon, XCircleIcon } from "@heroicons/react/24/outline";
 
 interface ValidationCheckProps {
@@ -6,9 +6,16 @@ interface ValidationCheckProps {
   message: string;
 }
 
-const ValidationCheck: React.FC<ValidationCheckProps> = ({ isValid, message }) => {
+const ValidationCheck: React.FC<ValidationCheckProps> = ({
+  isValid,
+  message,
+}) => {
   return (
-    <li className={`flex items-center mb-2 font-semibold ${isValid ? "text-green-700" : "text-red-700"}`}>
+    <li
+      className={`flex items-center mb-2 font-semibold ${
+        isValid ? "text-green-700" : "text-red"
+      }`}
+    >
       <div className="flex-shrink-0">
         {isValid ? (
           <CheckCircleIcon className="h-6 w-6 mr-3" />
@@ -25,9 +32,11 @@ interface ValidationChecklistProps {
   checks: ValidationCheckProps[];
 }
 
-const ValidationChecklist: React.FC<ValidationChecklistProps> = ({ checks }) => {
+const ValidationChecklist: React.FC<ValidationChecklistProps> = ({
+  checks,
+}) => {
   return (
-    <div className="p-2 -m-2 mt-6 mb-6">
+    <div className="p-2 -m-2">
       <ul>
         {checks.map((check, index) => (
           <ValidationCheck key={index} {...check} />
